Mac: Allow installing BOINC without installing Rosetta 2
